+++ src/Request.php
@@ -33,9 +33,27 @@ class Request
      */
     public function url(): string
     {
-        return $this->httpRequest()
+        $url = $this->httpRequest()
             ? $this->request->getUri()
             : '';
+
+        if (! $url) {
+            return $url;
+        }
+
+        // Manually filter out sensitive data from URL query string
+        $queryString = parse_url($url, PHP_URL_QUERY);
+        $filteredQueryParams = array_map(function ($keyAndValue) {
+            $parts = explode('=', $keyAndValue);
+            if (isset($parts[1]) && $parts[1] !== ''
+                && in_array($parts[0], $this->keysToFilter)) {
+                return "{$parts[0]}=[FILTERED]";
+            }
+
+            return $keyAndValue;
+        }, explode('&', $queryString));
+
+        return str_replace($queryString, implode('&', $filteredQueryParams), $url);
     }
 
     /**